requirements.txt
asgiref==3.7.1
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
